Costly chess - Dijkstras
In the country of Rome, Chess is a royal game. For evey move the players had to give some bucks to the Emperor Jurg. The LGMs or Little Green Men, are very good player of chess. But as the chess is a expensive game, thats why it is royal, they asked you to help them find the minimum bucks which they had to pay for moving their Knight from one position to another. Any number of steps can be used to reach the destination.

Constraints:

The chess has a dimension of 8X8, and the index of left bottom cell (0, 0).

Knight move only in a standard way, i.e. 2 row and 1 col or 1 row and 2 col.

If in a step Knight move from (a, b) to (c, d), then LGM had to pay a*c + b*d bucks to Emperor Jurg.

0 ≤ a, b, c, d ≤ 7

Input
There are 100-150 test cases. Each test case is composed of four space separeated integers.The first two numbers, a, b, are the starting position of the Knight and the next two, c, d, are the destination of the Knight. Read upto End Of File.

Output
For each test case, print the minimum amount of bucks they had to pay in separate line. If its impossible to reach the destination then print -1.

Example
Input:
2 5 5 2
4 7 3 2
1 2 3 4

Output:
42
78
18


Explanation for test case #1:
2 5 5 2

For moving Knight from (2, 5) to (5, 2) in minimum cost,  one of the path is (2, 5) -> (3, 3) ->(5, 2)
Bucks paid:
(2, 5)              =  0
(2, 5) -> (3, 3) =  0 + (2*3 + 5*3) = 21
(3, 3) -> (5, 2) = 21 + (3*5 + 3*2) = 42